Drop stale registrations without reducing CGC (#7594)
Currently the validator effective balance used for computing PeerDAS custody group count is only updated when the validator subscribes to the BN via  `validator/beacon_committee_subscriptions`.

If a validator stops registering with the node, the effective balance gets outdated and stays in the BN memory until the next restart. They are no longer required for CGC computation, as long as the CGC never reduces as per the spec, therefore they can be dropped.