Fix `DataColumnsByRoot` request limit validation bug (#7928)
Fixes #7926

This was a bug I introduced in #7890 and @pawanjay176 noticed it on some running nodes, and added a rpc test to confirm it.

The culprit is this line, where I failed to fill the vec to it's max size, so it doesn't calculate the max size properly, resulting in all `DataColumnByRoot` requests exceeding the max size during validation:
https://github.com/sigp/lighthouse/blob/d24a6d2a45f25dfdba8740d10c569226555a8143/consensus/types/src/chain_spec.rs#L1984

The PR fixes this and includes new regression tests for this fix.