Reduce noise in `Debug` impl of `RuntimeVariableList` (#8007)
The default debug output of these types contains a lot of unnecessary noise making it hard to read.

This PR removes the type and extra fields from debug output to make logs easier to read.

`len` could be potentially useful in some cases, but this gives us flexibility to only log it separately if we need it.

Related PR in `ssz_types`:
- https://github.com/sigp/ssz_types/pull/57


  


Co-Authored-By: Jimmy Chen <jchen.tc@gmail.com>